requirements.txt
Django==4.2.7
django-environ==0.11.2
requests==2.31.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
PyJWT==2.8.0
Authlib==1.2.1


